requirements.txt
# Core data processing
polars==1.9.0
duckdb==1.1.0

# dbt and database adapters
dbt-core==1.10.8
dbt-duckdb==1.10.0

# Orchestration
prefect==2.20.0
prefect-docker==0.5.0

# API and web requests
requests==2.32.0
httpx==0.27.0

# Music APIs
musicbrainzngs==0.7.1

# Data validation and quality
great-expectations==0.18.0

# Logging and monitoring
structlog==24.1.0
rich==13.7.0

# Configuration
python-dotenv==1.0.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Utilities
pandas==2.2.0  # for compatibility with some libraries
numpy==1.26.0
python-dateutil==2.9.0

# Development and testing
pytest==8.3.0
pytest-asyncio==0.23.0

# Optional: For future AWS integration
# boto3==1.35.0
# awscli==1.35.0